完成place_order服务逻辑

+++ app/order/biz/model/order_item.go
@@ -7,7 +7,7 @@ type OrderItem struct {
 	ProductId    uint32  `gorm:"type:int(11)"`
 	OrderIdRefer string  `gorm:"type:varchar(100);index"`
 	Quantity     int32   `gorm:"type:int(11)"`
-	Cost         float64 `gorm:"type:decimal(10,2)"`
+	Cost         float32 `gorm:"type:decimal(10,2)"`
 }
 
 func (OrderItem) TableName() string {
